+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.5.1 (tbd)
 
+- Fixed Node.js engine constraint in `piral-cli` to "16.0.0"
+- Updated `kras` to latest `v0.16` release
 - Added `--type` flag to `piral publish` command
 
 ## 1.5.0 (February 25, 2024)