+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed Node.js engine constraint in `piral-cli` to "16.0.0"
 - Fixed `target`-specific navigation in `piral-blazor`
+- Fixed definition of not found page for React Router v6 (#677)
 - Updated `kras` to latest `v0.16` release
 - Added `--type` flag to `piral publish` command